Update README.md
Internship Projects by Feynn Lab
Project 1: Ideation Project
This project involves searching for a topic and exploring how data science and machine learning can be applied to it and also required to prepare a comprehensive strategy based on our findings.
Project 2: Market Segmentation and EDA
In this project, we will provide a summary of the fundamentals of market segmentation. Additionally, we will perform exploratory data analysis (EDA) on a McDonald's dataset.